Resume_Q&A/requirements.txt
pdf2image
google-generativeai
python-dotenv
streamlit